requirements.txt
# API Base
fastapi[standard]~=0.115
pydantic-settings~=2.7
SQLAlchemy~=2.0
asyncpg~=0.30
setuptools-scm~=8.1
pyyaml~=6.0

# CDKTF
cdktf>=0.20
cdktf-cdktf-provider-aws>=19.60

# Auth
pyjwt
bcrypt
boto3>=1.38.0

# Crypto
cryptography>=41.0.0
argon2-cffi>=23.1.0

# Async Jobs
arq>=0.26.3
aiofiles>=24.1.0
tenacity>=9.1.2